+++ src/main/java/com/svalero/enajenarte/service/UserService.java
@@ -2,19 +2,25 @@ package com.svalero.enajenarte.service;
 
 import com.svalero.enajenarte.domain.User;
 import com.svalero.enajenarte.domain.Registration;
-import com.svalero.enajenarte.domain.enums.PaymentStatus;
+import com.svalero.enajenarte.domain.ProgramRegistration;
+import com.svalero.enajenarte.dto.ProgramRegistrationOutDto;
+import com.svalero.enajenarte.dto.UserEditInDto;
 import com.svalero.enajenarte.dto.UserInDto;
 import com.svalero.enajenarte.dto.UserOutDto;
 import com.svalero.enajenarte.dto.UserRegistrationOutDto;
+import com.svalero.enajenarte.exception.AccessDeniedException;
+import com.svalero.enajenarte.exception.HasAssociatedRegistrationsException;
 import com.svalero.enajenarte.exception.UserNotFoundException;
 import com.svalero.enajenarte.repository.UserRepository;
 import com.svalero.enajenarte.repository.RegistrationRepository;
+import com.svalero.enajenarte.repository.ProgramRegistrationRepository;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.security.core.context.SecurityContextHolder;
 
-import java.time.LocalDate;
 import java.util.List;
 import java.util.ArrayList;
 
@@ -26,7 +32,11 @@ public class UserService {
     @Autowired
     private RegistrationRepository registrationRepository;
     @Autowired
+    private ProgramRegistrationRepository programRegistrationRepository;
+    @Autowired
     private ModelMapper modelMapper;
+    @Autowired
+    private PasswordEncoder passwordEncoder;
 
 
     // GET all
@@ -38,7 +48,7 @@ public class UserService {
         final Boolean finalActive = active.isEmpty() ? null : Boolean.parseBoolean(active);
 
         // filtrado con stream
-        List<User> filteredusers = userRepository.findAll().stream()
+        List<User> filteredUsers = userRepository.findAll().stream()
                 .filter(user -> finalUsername == null || user.getUsername().toLowerCase().contains(finalUsername))
                 .filter(user -> finalEmail == null || user.getEmail().toLowerCase().contains(finalEmail))
                 .filter(user -> finalActive == null || user.isActive() == finalActive)
@@ -46,8 +56,19 @@ public class UserService {
 
         // Mapear DTOs
         List<UserOutDto> userOutDtoList =
-                modelMapper.map(filteredusers, new TypeToken<List<UserOutDto>>() {
-                }.getType());
+                modelMapper.map(filteredUsers, new TypeToken<List<UserOutDto>>() {}.getType());
+
+        for (int i = 0; i < filteredUsers.size(); i++) {
+            User user = filteredUsers.get(i);
+            UserOutDto dto = userOutDtoList.get(i);
+
+            if (user.getGender() != null) {
+                dto.setGender(user.getGender().getDisplayName());
+            }
+            if (user.getAgeGroup() != null) {
+                dto.setAgeGroup(user.getAgeGroup().getDisplayName());
+            }
+        }
 
         return userOutDtoList;
     }
@@ -56,12 +77,65 @@ public class UserService {
     public UserOutDto findById(long id) throws UserNotFoundException {
         User user = userRepository.findById(id)
                 .orElseThrow(UserNotFoundException::new);
-        return modelMapper.map(user, UserOutDto.class);
+
+        UserOutDto userOutDto = modelMapper.map(user, UserOutDto.class);
+
+        if (user.getGender() != null) {
+            userOutDto.setGender(user.getGender().getDisplayName());
+        }
+        if (user.getAgeGroup() != null) {
+            userOutDto.setAgeGroup(user.getAgeGroup().getDisplayName());
+        }
+        return userOutDto;
     }
 
-    public List<UserRegistrationOutDto> getUserRegistrations(long userId) throws UserNotFoundException {
+    public UserOutDto findCurrentUser() throws UserNotFoundException {
+        String authenticatedUsername = SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getName();
+
+        User user = userRepository.findByUsername(authenticatedUsername);
+
+        if (user == null) {
+            throw new UserNotFoundException();
+        }
+
+        UserOutDto userOutDto = modelMapper.map(user, UserOutDto.class);
+
+        if (user.getGender() != null) {
+            userOutDto.setGender(user.getGender().getDisplayName());
+        }
+        if (user.getAgeGroup() != null) {
+            userOutDto.setAgeGroup(user.getAgeGroup().getDisplayName());
+        }
+
+        return userOutDto;
+    }
+
+    // Obtener las inscripciones del usuario
+    public List<UserRegistrationOutDto> getUserRegistrations(long userId) throws UserNotFoundException, AccessDeniedException {
         User user = userRepository.findById(userId)
                 .orElseThrow(UserNotFoundException::new);
+
+        // Usuario autenticado
+        String authenticatedUsername = org.springframework.security.core.context.SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getName();
+
+        String authenticatedRole = org.springframework.security.core.context.SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getAuthorities()
+                .iterator()
+                .next()
+                .getAuthority();
+
+        // Solo el propio usuario o ADMIN
+        if (!user.getUsername().equals(authenticatedUsername) && !authenticatedRole.equals("ROLE_ADMIN")) {
+            throw new AccessDeniedException();
+        }
 
         List<Registration> registrations = registrationRepository.findByUser(user);
         List<UserRegistrationOutDto> userRegistrationOutDtos = new ArrayList<>();
@@ -84,24 +158,76 @@ public class UserService {
 
             userRegistrationOutDtos.add(userRegistrationOutDto);
         }
+
         return userRegistrationOutDtos;
+    }
+
+    public List<ProgramRegistrationOutDto> getUserProgramRegistrations(long userId)
+            throws UserNotFoundException, AccessDeniedException {
+
+        User user = userRepository.findById(userId)
+                .orElseThrow(UserNotFoundException::new);
+
+        String authenticatedUsername = SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getName();
+
+        String authenticatedRole = SecurityContextHolder
+                .getContext()
+                .getAuthentication()
+                .getAuthorities()
+                .iterator()
+                .next()
+                .getAuthority();
+
+        if (!user.getUsername().equals(authenticatedUsername) && !authenticatedRole.equals("ROLE_ADMIN")) {
+            throw new AccessDeniedException();
+        }
+
+        List<ProgramRegistration> registrations = programRegistrationRepository.findByUser(user);
+        List<ProgramRegistrationOutDto> outDtos = new ArrayList<>();
+
+        for (ProgramRegistration registration : registrations) {
+            ProgramRegistrationOutDto programRegistrationOutDto = modelMapper.map(registration, ProgramRegistrationOutDto.class);
+
+            programRegistrationOutDto.setFullName(user.getFullName());
+            programRegistrationOutDto.setUserId(registration.getUser().getId());
+            programRegistrationOutDto.setProgramName(registration.getProgram().getName());
+            programRegistrationOutDto.setProgramId(registration.getProgram().getId());
+
+            outDtos.add(programRegistrationOutDto);
+        }
+
+        return outDtos;
     }
 
     // POST
     public UserOutDto add(UserInDto userInDto) {
         User user = modelMapper.map(userInDto, User.class);
 
+        user.setPassword(passwordEncoder.encode(userInDto.getPassword()));
+
         // generadas por el sistema
         user.setRole("USER");
         user.setActive(true);
-        user.setBalance(0);
 
         User newUser = userRepository.save(user);
-        return modelMapper.map(newUser, UserOutDto.class);
+
+        UserOutDto userOutDto = modelMapper.map(newUser, UserOutDto.class);
+
+        if (newUser.getGender() != null) {
+            userOutDto.setGender(newUser.getGender().getDisplayName());
+        }
+        if (newUser.getAgeGroup() != null) {
+            userOutDto.setAgeGroup(newUser.getAgeGroup().getDisplayName());
+        }
+
+        return userOutDto;
     }
 
     // PUT
-    public UserOutDto modify(long id, UserInDto userInDto) throws UserNotFoundException {
+    public UserOutDto modify(long id, UserEditInDto userEditInDto) throws UserNotFoundException, AccessDeniedException {
         User existingUser = userRepository.findById(id)
                 .orElseThrow(UserNotFoundException::new);
 
@@ -122,30 +248,50 @@ public class UserService {
                     .getAuthority();
 
             if (!role.equals("ROLE_ADMIN")) {
-                throw new RuntimeException("You cannot modify another user");
+                throw new AccessDeniedException();
             }
         }
 
-        // Datos de sistema
-        String role = existingUser.getRole();
-        boolean active = existingUser.isActive();
-        float balance = existingUser.getBalance();
+        existingUser.setEmail(userEditInDto.getEmail());
+        existingUser.setFullName(userEditInDto.getFullName());
+        existingUser.setPhone(userEditInDto.getPhone());
+        existingUser.setGender(userEditInDto.getGender());
+        existingUser.setAgeGroup(userEditInDto.getAgeGroup());
 
-        modelMapper.map(userInDto, existingUser);
-        existingUser.setId(id);
+        if (userEditInDto.getPassword() != null
+                && !userEditInDto.getPassword().isBlank()) {
+            existingUser.setPassword(passwordEncoder.encode(userEditInDto.getPassword()));
+        }
 
-        existingUser.setRole(role);
-        existingUser.setActive(active);
-        existingUser.setBalance(balance);
+        User updatedUser = userRepository.save(existingUser);
 
-        User updateUser = userRepository.save(existingUser);
-        return modelMapper.map(updateUser, UserOutDto.class);
+        UserOutDto userOutDto = modelMapper.map(updatedUser, UserOutDto.class);
+
+        if (updatedUser.getGender() != null) {
+            userOutDto.setGender(updatedUser.getGender().getDisplayName());
+        }
+        if (updatedUser.getAgeGroup() != null) {
+            userOutDto.setAgeGroup(updatedUser.getAgeGroup().getDisplayName());
+        }
+
+        return userOutDto;
     }
 
     // DELETE
-    public void delete(long id) throws UserNotFoundException {
+    // No se puede eliminar un usuario con inscripciones asociadas
+    public void delete(long id) throws UserNotFoundException, HasAssociatedRegistrationsException {
         User user = userRepository.findById(id)
                 .orElseThrow(UserNotFoundException::new);
+
+        List<Registration> registrations = registrationRepository.findByUser(user);
+        if (!registrations.isEmpty()) {
+            throw new HasAssociatedRegistrationsException();
+        }
+
+        List<ProgramRegistration> programRegistrations = programRegistrationRepository.findByUser(user);
+        if (!programRegistrations.isEmpty()) {
+            throw new HasAssociatedRegistrationsException();
+        }
 
         userRepository.delete(user);
     }
@@ -158,9 +304,26 @@ public class UserService {
             throw new UserNotFoundException();
         }
 
-        if (!user.getPassword().equals(password)) {
+        String storedPassword = user.getPassword();
+
+        // comprueba si el usuario ya ha sido migrado a BCrypt
+        if (storedPassword != null && storedPassword.startsWith("$2")) {
+            if (!passwordEncoder.matches(password, storedPassword)) {
+                throw new UserNotFoundException();
+            }
+
+            return user;
+        }
+
+        // Compatibilidad temporal con los usuarios existentes:
+        if (storedPassword == null || !storedPassword.equals(password)) {
             throw new UserNotFoundException();
         }
+
+        // Si el usuario antiguo se autentica correctamente, su contraseña se migra automáticamente a BCrypt
+        user.setPassword(passwordEncoder.encode(password));
+        userRepository.save(user);
+
         return user;
     }
 }
